+++ src/mod/LaminaPeek.h
@@ -1,6 +1,15 @@
 #pragma once
 
+#include "ll/api/event/ListenerBase.h"
 #include "ll/api/mod/NativeMod.h"
+
+#include "mod/Config.h"
+#include "mod/preview/HoveredPreviewCache.h"
+#include "mod/render/PreviewRenderer.h"
+
+namespace ll::event::inline render {
+class AfterUIRenderEvent;
+}
 
 namespace lamina_peek {
 
@@ -13,6 +22,8 @@ public:
 
     [[nodiscard]] ll::mod::NativeMod& getSelf() const { return mSelf; }
 
+    [[nodiscard]] Config const& getConfig() const { return mConfig; }
+
     /// @return True if the mod is loaded successfully.
     bool load();
 
@@ -23,7 +34,13 @@ public:
     bool disable();
 
 private:
-    ll::mod::NativeMod& mSelf;
+    void onAfterUIRender(ll::event::AfterUIRenderEvent& event);
+
+    ll::mod::NativeMod&          mSelf;
+    Config                       mConfig;
+    ll::event::ListenerPtr       mUIRenderListener;
+    preview::HoveredPreviewCache mPreviewCache;
+    render::PreviewRenderer      mPreviewRenderer;
 };
 
 } // namespace lamina_peek